moved template routes to src lib; start footer
