+++ src/lib/svelteTemplates/Footer.svelte
@@ -1,0 +1,18 @@
+<footer>
+	<div class='footer-header'>Other Pages</div>
+	<div class='footer-header'>Other Resources</div>
+	<a href='https://b2popology.com/'>Battles 2 Popology</a>
+	<a href='/updates'>Website Updates</a>
+</footer>
+
+<style>
+	footer {
+	  display: grid;
+	  grid-template-columns: auto auto;
+	  background-color: #f0f0f0; /* Add a background color for the container */
+	  justify-items: center;
+	}
+	.footer-header {
+		font-family: Montserrat-SemiBold, Arial, Helvetica, sans-serif;
+	}
+</style>
