+++ src/lib/svelteTemplates/TableStatsData.svelte
