+++ src/lib/svelteTemplates/TableStatsDataSmall.svelte
