+++ src/lib/svelteTemplates/TableStatsHeader.svelte
