+++ src/lib/svelteTemplates/TableStatsHeaderSmall.svelte
