+++ src/lib/svelteTemplates/TableStatsWhole.svelte
