+++ src/lib/svelteTemplates/TableStatsWholeSmall.svelte
