+++ src/routes/heroes/+page.svelte
@@ -1,7 +1,7 @@
 <script>
-  import FilterDropdown from "./../FilterDropdown.svelte"
-  import Nav from "./../Nav.svelte"
-  import TableStatsWhole from "./../TableStatsWhole.svelte"
+  import FilterDropdown from "$lib/svelteTemplates/FilterDropdown.svelte"
+  import Nav from "$lib/svelteTemplates/Nav.svelte"
+  import TableStatsWhole from "$lib/svelteTemplates/TableStatsWhole.svelte"
   import { mapNames, heroNames } from "$lib/ThingAliases"
   let selectedMap = ""
